requirements.txt
streamlit=1.54.0
msal=1.34.0
pandas=2.3.3
openpyxl=3.1.5
python-dotenv=1.2.1
requests=2.32.5